Implementación del test BeAbleToAddItsEarningsToExistingEarningsValueGivenFromAndToDates() en la clase EarningQueriesServiceShould. Estado: No compila.

+++ TodoPagos/TodoPagos.Web.Services/EarningQueriesService.cs
@@ -23,7 +23,7 @@ namespace TodoPagos.Web.Services
             if (oneUnitOfWork == null) throw new ArgumentException();
         }
 
-        public int GetAllEarnings(DateTime from, DateTime to)
+        public double GetAllEarnings(DateTime from, DateTime to)
         {
             throw new NotImplementedException();
         }
